Rename module to 'constants.py' to make its purpose clearer
